Replaced the Photon in the simulation with VanillaPhoton.  It works except that the photon list just keeps getting bigger, which means it's failing to remove from collections when destroyed.  I would think that's because I don't use events to alert the collection when an object is destroyed, and the photons are shoved into multiple collections.  I need to refactor so that the `bodies` collection isn't a collection (or isn't a Backbone.Collection, because it means the photons being added are just wrapped in Backbone models, which defeats the purpose) and so the VanillaCollection makes the objects keep track of any number of collections.
